requirements.txt
flask>=3.0
anthropic>=0.40
requests>=2.31
python-dotenv>=1.0
urllib3>=2.0
matplotlib>=3.8
sqlalchemy>=2.0
